Add password visibility toggle to AddEditActivity
This commit introduces a password visibility toggle feature on the Add/Edit screen.

A `setOnCheckedChangeListener` is added to the `switchShowPassword` widget. When toggled, it changes the `transformationMethod` of the password `EditText` between `PasswordTransformationMethod` and `HideReturnsTransformationMethod` to show or hide the password text. The cursor is also moved to the end of the text after toggling.